Remove the default service
This change removes the default service as we have with other routers,
moving the default target group into the main.tf file. Some time after
this has merged (i.e. when it has been applied to all components that
use it), we should be able to remove the modules/deprecated folder (the
default target group must be replaced before we can delete the old target
group).