Dedupe the default target group name
Stupid terraform evaluates both sides of a ternary operator, which makes
this a bit worse looking than it would otherwise be. If the
env-component is 32 chars or less use that, otherwise truncate it to 24
chars and use the first 8 chars of the sha1(env-component) to make up
the rest - should avoid duplicates.