Abbreviate both ALB and default target group names
The only service using this in live is the document-repository-router.
This doesn't use the default target group yet, and by happy coincidence
has the ALB name document-repository-router - hence it should be safe to
move to the new naming scheme by just breaking things in aslive. It was
necessary to remove component names beginning or ending with a hyphen
since tf_alb currently strips these off.